% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcreplicate.R
\name{mcreplicate}
\alias{mcreplicate}
\title{Multi-core replicate. From the rethinking package:
<https://github.com/rmcelreath/rethinking/blob/3b48ec8dfda4840b9dce096d0cb9406589ef7923/R/utilities.r#L206>}
\usage{
mcreplicate(n, expr, refresh = 0.1, mc.cores = 2)
}
\arguments{
\item{n}{integer: the number of replications.}

\item{expr}{the expression (a language object, usually a call) to evaluate repeatedly.}

\item{refresh}{status update refresh interval}

\item{mc.cores}{number of cores to use}
}
\description{
Multi-core replicate. From the rethinking package:
<https://github.com/rmcelreath/rethinking/blob/3b48ec8dfda4840b9dce096d0cb9406589ef7923/R/utilities.r#L206>
}
